Supporting dark mode on macOS and iOS
Also fixed some bugs when using on UIView and NSView.

Got it to work for SwiftUI on macOS. SwiftUI on iOS seems borked beyond all hope. It will continually ask for intrinsic content size without ever setting the frame. It sets the frame at the end to the actual width, but invalidating the intrinsic size does nothing.

Mojave is lacking in full dark mode support, because Apple didn't see the need to provide the necessary APIs.