Adding SwiftUI bindings
# Problem

SwiftUI is the new hotness, and it'd be nice to have a SwiftUI wrapper for NativeMarkLabel.

The problem is [SwiftUI doesn't respect UIViewRepresentables invalidating their intrinsicContentSize](https://losingfight.com/blog/2020/08/22/uiviewrepresentable-doesnt-respect-intrinsiccontentsize-invalidation/).

# Solution

Apply some hacks/workarounds to properly adjust the height of `NativeMarkText`. First, add an internal callback on `NativeMarkLabel` so we can know when the intrinsic content size does update. Use a `Binding` to feed that size back up into a SwiftUI `View` wrapping our `UIViewRepresentable`. This allows us to use the `.frame(height:)` SwiftUI modifier to adjust the correct height correctly.

The other bit of hackery is dynamically applying the horizontal hugging priority to `NativeMarkLabel` based on if the text is multiline or not. If it's multiline we don't force horizontal hugging, but if it's single line we do. If we force hugging on multiline we create a bug, most noticeable on macOS. If we resize a window containing a `NativeMarkText` smaller then larger, this causes the `NativeMarkText` to keep the smallest size it's ever been, leading to an ever-shrinking text.