NativeMarkText doesn't build on armv7 [#3]
# Problem

As reported in #3, the NativeMarkText+UIKit doesn't build for device. Investigation showed that was because the enclosing projects were trying to build armv7 and SwiftUI doesn't include an armv7 binary. Therefore, it appeared like SwiftUI wasn't defined and compiling the file failed with undefined symbols.

# Solution

As a workaround, only build the NativeMarkText+UIKit.swift file for armv64 and x86_64. It is a bit fragile in that if new architectures appear, they'll have to be manually added.