Fix compiler crash building i386 [#7] (#8)
* Fix compiler crash building i386 [#7]

# Problem

When building release configuration of certain iOS apps, Xcode will ask for i386 simulator builds for NativeMarkKit. This happens if the Xcode project doesn't use these old versions of iOS. The problem is with Xcode 12.5 there is a crashing compiler bug when building NativeMarkKit for i386.

# Solution

Bump the minimum iOS and tvOS versions to 11. At this point Xcode 12.5 no longer attempts to build any i386 Simulators and we are spared compiler crashes.

* Replaced the deprecated `openURL:` method with its new counter part for iOS10+ (#9)

Co-authored-by: Andrew Marinov <>

Co-authored-by: andrewsexcellent <82413230+andrewsexcellent@users.noreply.github.com>