Fix italic not preserving weight on custom fonts [#10]
#10

# Problem

For custom fonts where a non-default weight has been specified (e.g. medium or roman), adding italic causes that weigh to be lost and the font will go back to the "regular" weight.

# Solution

On AppKit, use `NSFontManager.convert(_:toHaveTrait:)` which correctly maintains the weight.

On UIKit, we have to manage things manually. If we have already specified bold or about to, then we want the change in weight, so we use the usual `withSymbolicTraits`. However, if we're just adding italic, we want to preserve the weight. To do that, we manually fetch all the fonts of the given family, match against the symbolic traits (bold, italic, monospace) and the weight. We pick the closest font we can find.