Fix large image rendering bugs
# Problem

On macOS in particular, large image attachments were not rendering at all or not rendering until the window was resizes. This was because of a couple of issues:

1. The full affected range was not being invalidated
1. If the size of the image forced it outside the container, no view-level invalidation happened
1. The view didn't invalidate its intrinsic content size when an image was loaded

# Solution

Capture and use the full affected character range, provide a brute force invalidation fallback of the image is forced outside the container, and invalid the intrinsic size on image load.