Support transparent background for NativeMarkText
# Problem

It is not currently possible to set the background of a `NativeMarkText` or `NativeMarkLabel` to be completely transparent.

# Solution

First, allow the `.backgroundColor` to take an optional `NativeColor`. This means a user can specify:

```Swift
.document: [
     .backgroundColor(nil)
]
```
to indicate they don't want a background color. This actually fixes it for macOS (i.e. AppKit).

However, UIKit is a special case. For it to work, the `backgroundColor` of the `UIView` _also_ has to be set to `UIColor.clear` otherwise the background still ends up solid black. To fix this, add special code to pull the document's background color and check if its `nil`. If it is `nil`, set `backgroundColor = .clear`.